Unittest fixes
Reading CSV treats integer columns with nulls as floats causing random
when a null value is a chunk (treating id as float)  or not (treating
as an int). This causes errors writing the dataset occasionally
depending on where we randomly inserted the null value.

Also fix so that we read in at least one row group, and handle case
where the dropna op doesn't drop any rows.