Change ZeroFill to FillMissing / Clip (#203)
Rename ZeroFill to Clip/FillMissing

The ZeroFill operator was not only replacing missing values with 0, it
was also replacing negative values with 0. Since this can cause confusion
this change modifies ZeroFill to a Clip operator.

The 'FillMissing' operator already handles the case of replacing missing
values to 0 so we don't need a replacement for that functionality.