Fixes for Filter op (#329)
The Filter op had a bad example in the docstring: this op required the
Filter function to return a DataFrame - but the docstring was
return a series of bool's indicating rows that should be filtered out.
Fix by supporting Filter ops returning a series of bools, as well
as a filtered DataFrame. Also when the Filter op returns
invalid types return an error right away rather than failing
downstream in the workflow.