API Overhaul (#491)
This change extends our preprocessing and feature engineering workflows to be directed graphs of operators applied onto user defined groups of columns - and not just represent workflows as lists of operators applied to only categorical and continuous columns.  Defining  this graph is decoupled from the Workflow class, and which lets users easily define complicated graphs of operations on their own custom defined sets of columns. 

This also changes our Workflow to adopt a similar API to that found with transformers in sklearn: statistics are calculated with a ‘fit’ method and applied with a ‘transform’ method. The NVTabular Dataset object is extended to handle both input and output of datasets,  with the ‘transform’ method of the workflow both taking an input Dataset and returning as output a transformed Dataset. Likewise each Operator also uses a sklearn transformer API, and is responsible for owning its own statistics.