Allow nested groups to use other columngroups (#525)
When doing combo categorify or targetencoding, there are cases where we want the
input to be on groups that include columns that are the output of other operators.
Previously we only allowed groups of untransformed columns. This updates the ColumnGroup
class to allow groups that contain other ColumnGroups as one of the elements of the group.

Closes #522