Workflow serialization (#536)
The existing load_stats/save_stats code only saved the calculated statistics for a workflow, but
didn't save all the operators that were used as well as the processing graph. This makes it
difficult to recreate the workflow exactly for cases like inferences

This PR adds support for saving and loading the entire workflow object. Since the operators
being applied could use lambdas or be custom operators written in jupyter notebooks - we're
using cloudpickle for serialization (cloudpickle is also used by dask).