[FEA] PyTorch Dataloader support for multi-GPU training w/ Horovod (#674)
* PyTorch Dataloader support for multi-GPU training w/ Horovod

This is a working training script for the Movielens example with multiple GPUs but no data partitioning. (All workers train on exactly the same data, since the dataloader itself doesn't yet have support for partitioning the data across workers.)

* Add dataset partitioning for multi-GPU into the PyTorch dataloader

* Use `horovod.join()` to handle workers with uneven number of batches

`join()` blocks until all Horovod workers arrive at the same point, and in the mean time, it submits the identity value (e.g. zeros) to any allgather/allreduce operations. This allows whichever workers have the most data to process to go ahead and process it without being blocked by the workers with less data.

* Use all Horovod workers to perform validation

This requires broadcasting the model params again after each training epoch, so that models that had less data and were previously blocked on the `join()` can catch their models up to the latest params before validating.

* Generate consistent shuffle seeds across workers with Horovod

* Only reseed before shuffle if a seeding function exists

This makes the shuffles reproducible by avoiding a step inside the dataloader that reseeds with system randomness. You can still do that in a seeding function if you want, but this allows you to set a single static seed up front and get the same series of shuffles again.