Fix for outbrain TF inference
Previously we couldn't export the outbrain model to be serverd up with triton.
One issue was that the W&D model we are using doesn't have inputs/outputs
defined on it, and we failed to generate the triton config because of this.
Fix by falling back to using the default_serving signature in this case.

Another issue was that the W&D model has two optimizers. Fix by not saving
both optimizers when we save the model.