Fit a single stat operator at a time
We're seeing unittest failures, usually along the lines of
``` TypeError: float() argument must be a string or a number, not '_NAType'```

This seems to only occur for me when calculating the stats for both FillMedian
and Categorify in a single dask.compute call. Calculating each op independently
doesn't have this problem and seems to work consistently.