Fix hugectr notebooks
The recent changes for black-nb and flake8-nb messed up the hugectr notebooks.

Specifically applying black-nb against JSON formatted data added a trailing comma
to most entries, meaning it couldn't be parsed as JSON. Also we need the MPI import
for hugectr training.

Fix - and disable black-nb until it can be disabled on a per-cell level