Fix partition length calculation (#795)
We're seeing some failures in the TF movielens notebooks, failing with

```
/nvtabular/nvtabular/io/dataset.py in <genexpr>(.0)
    913             # will not be correct if rows where added or dropped
    914             # after IO (within Ops).
--> 915             return sum(self.partition_lens[i] for i in self.indices)
    916         if len(self.indices) < self._ddf.npartitions:
    917             return len(self._ddf.partitions[self.indices])
IndexError: list index out of range
```

This error seems to be because we have only a single partition_lens available,
but there are two partitions and indices in the dataset.

Fix by fixing the logic that calculates this partition_lens. The row_groups_per_part
was being included twice in the iterating over the partitions (once in calculating the
number of partitions, and also once in the step size).