Provide better graph labels for LambaOp
The workflow graphs all had 'LambdaOp' before, which was pretty confusing if you
had multiple lambdas in the workflow. This change adds a 'label' to the operator
api which is used to label nodes in the graph.

LambdaOp will now allow users to specify their own label, or alternatively will
try to figure out a decent representation from the function itself using the
inspect module.