Add 'supports' and 'inference_initialize' for operators. (#918)
This is the first in a series of PR's for providing optimized inference.

This adds both a 'supports' flag for operators and a 'inference_initialize' method.

The supports flag shows what dataformats the operator can accept during transforms.
We've found for inference that a dictionary of column names to numpy/cupy arrays
can be orders of magnitude faster at inference time. As an example, the normalize operator
works unchanged on these dictionaries, and times for me go from 8ms to around 140us on the
Rossmann dataset using dictionaries of numpy arrays instead of cudf or pandas dataframes.

The inference_initialize change provides a mechanism for operators to provide an optimized
version for inference, as well as to preload any necessary data. This is currently only
serving to preload categories in the categorify op - but will be used to provide optimize
versions of operators in a future PR.