Update triton model.py to use dict array representations
Update the triton model.py to use dict representations of operators when allowed
(CPU_DICT_ARRAY or GPU_DICT_ARRAY).  This allows us to avoid representing
data in a dataframe format entirely where possible and enables future optimizations.
This is the 2nd of 3 PR's to enable small batch optimizations for inference.