Fix installing with pip
Installing with pip fails because of a 'versioneer' not found error. This
error seems to be related to using a newer version of pip alongside with the
build_system in pyproject.toml from pep 517/518.

Fix by adding the current directory to the python path as suggested in
https://github.com/python-versioneer/python-versioneer/issues/193 if the
versioneer import fails

Closes https://github.com/NVIDIA/NVTabular/issues/973