Convert `ColumnSelector` to a class (instead of a type alias)
Eventually, we're going to `ColumnSelector` to be able to select by tag, which is going to require it to implement some methods. A type alias can't have methods, so we need it to be an actual class. For now, let's use a dataclass and add the minimum number of methods we need to make the tests pass again.