Add missing calls to `ColumnSelector.names` and fix `Workflow`
This commit adds additional calls to `col_selector.names` that weren't necessary to make the tests pass, but are clearly needed from a visual inspection of the call sites. Those changes in turn break some tests related to the `Workflow` class, which still needed a few adjustments for the introduction of `ColumnSelector`.