Make left side of `>>` always a `ColumnSelector` or `ColumnGroup`
This changes the supported types of `>>` expressions so that the left side is always an NVT object (and never a Python primitive type like a `list`.) This lets us define `>>` on all of the allowed left side types, which then lets us remove the definition of `Operator.__rrshift`, which removes the circular dependency between `Operator` and `ColumnGroup`.

The circular dependency shifts to `Operator` and `ColumnSelector`, but this is still an improvement, because now `Operator` only depends on the column selecting behavior that used to be in `ColumnGroup` and not the workflow graph creating functionality that's still in `ColumnGroup`.

(There are still a few residual connections between `ColumnGroup` and `Operator` to be removed, but these require changing the behavior in `ColumnGroup.__init__`, which I'll save for a future commit since the tests are passing here.)