Restrict `ColumnGroup` constructor arg to be selector or group
Restricting the constructor argument type catches some additional cases where `ColumnGroup` was being used to represent a list of columns. The outstanding issue that prevents the tests from passing here is that there's not yet a good way to represent nested `ColumnSelectors` in the DSL for building up graphs. `ColumnSelector` is perfectly capable of being nested, but should also never contain a `ColumnGroup` (i.e. workflow node), which wasn't an issue when `ColumnGroup` was both the node and the selector.

To merge two lists of column names, we can use the `+` operator, but there's not yet an equivalent syntax to say "take this `ColumnGroup`, grab its `ColumnSelector`, make the list of names into a subgroup, and then merge it with this other `ColumnGroup`." Still pondering how to do that without allowing lists to be `>>`'d onto operators.