Restore the `["a", "b", "c"] >> ops.Operator()` syntax
Removing this helped separate `ColumnGroup` into `ColumnSelector` and `WorkflowNode` (the new name of the graph parts of `ColumnGroup` applied in a future PR), but it's still possible to implement this after the big refactor in order to maintain backward compatibility. This commit rolls back changes to the test suite and notebook related to the earlier removal of this syntax, re-implements the behavior, and adds a deprecation warning.