Introduce the concept of a `ColumnSelector` (#1014)
* Rename `ColumnNames` to `ColumnSelector`

With the addition of tagging, we're going to want to select columns from dataframes by their tags instead of their names. This refactor generalizes the name of the existing column selection type (`ColumnNames`) to account for the new use case (to be added in future commits.)

* Convert `ColumnSelector` to a class (instead of a type alias)

Eventually, we're going to `ColumnSelector` to be able to select by tag, which is going to require it to implement some methods. A type alias can't have methods, so we need it to be an actual class. For now, let's use a dataclass and add the minimum number of methods we need to make the tests pass again.

* Convert col names from `FitOptions` into `ColumnSelector` earlier

This moves the creation of the `ColumnSelector` object farther up in the `_top_level_groupby` function used by `Categorify` and adjusts the variable names to distinguish between the list of column names from `FitOptions` and the column selector object.

* Move conversion to `ColumnSelector` into `FitOptions`

Instead of converting to `ColumnSelector` inside `_top_level_groupby()`, this commit pushes the change farther out into the rest of the code base by moving the conversion into `FitOptions` itself and then adjusting the downstream code that breaks as a result.

* Add missing calls to `ColumnSelector.names` and fix `Workflow`

This commit adds additional calls to `col_selector.names` that weren't necessary to make the tests pass, but are clearly needed from a visual inspection of the call sites. Those changes in turn break some tests related to the `Workflow` class, which still needed a few adjustments for the introduction of `ColumnSelector`.

* Move `ColumnSelector` to its own file

* Enable nested `ColumnSelectors`

This commit adds functionality to the `ColumnSelector` class to allow nesting selectors, which simplifies the types and makes cases where grouped sets of columns are used explicit.

* Make left side of `>>` always a `ColumnSelector` or `ColumnGroup`

This changes the supported types of `>>` expressions so that the left side is always an NVT object (and never a Python primitive type like a `list`.) This lets us define `>>` on all of the allowed left side types, which then lets us remove the definition of `Operator.__rrshift`, which removes the circular dependency between `Operator` and `ColumnGroup`.

The circular dependency shifts to `Operator` and `ColumnSelector`, but this is still an improvement, because now `Operator` only depends on the column selecting behavior that used to be in `ColumnGroup` and not the workflow graph creating functionality that's still in `ColumnGroup`.

(There are still a few residual connections between `ColumnGroup` and `Operator` to be removed, but these require changing the behavior in `ColumnGroup.__init__`, which I'll save for a future commit since the tests are passing here.)

* Raise exceptions in `ColumnSelector` when an element is a `ColumnGroup`

These exceptions enforce the invariants about the relationships between `ColumnSelectors` and `ColumnGroups`. The proper relationships are that a `ColumnGroup` should have a `ColumnSelector` and may have other `ColumnGroups` as parents, but a `ColumnSelector` should never have or be the parent of a `ColumnGroup`.

* Restrict `ColumnGroup` constructor arg to be selector or group

Restricting the constructor argument type catches some additional cases where `ColumnGroup` was being used to represent a list of columns. The outstanding issue that prevents the tests from passing here is that there's not yet a good way to represent nested `ColumnSelectors` in the DSL for building up graphs. `ColumnSelector` is perfectly capable of being nested, but should also never contain a `ColumnGroup` (i.e. workflow node), which wasn't an issue when `ColumnGroup` was both the node and the selector.

To merge two lists of column names, we can use the `+` operator, but there's not yet an equivalent syntax to say "take this `ColumnGroup`, grab its `ColumnSelector`, make the list of names into a subgroup, and then merge it with this other `ColumnGroup`." Still pondering how to do that without allowing lists to be `>>`'d onto operators.

* Make adding a list of `ColumnGroups` to another turn them into subgroups

This mirrors the existing syntax, but requires that the list be on the right. I think we can fix that by adding a `__radd` method to `ColumnGroup`.

* Simplify `ColumnGroup` constructor

This converts the assertion (which was handy for development) into a `TypeError` and removes code that can no longer be reached since the `columns` parameter must always be a `ColumnSelector`.

* Update Torch/Horovod multi-GPU test to use `ColumnSelector`

* Update Triton inference tests to use `ColumnSelector`

* Update notebooks to use `ColumnSelector`

* Add `ColumnSelector` as a top-level export (for the notebooks)

We might be able to remove `ColumnGroup` from the exports there, but I'm going to leave it for now.

* Update Criteo notebook to use `ColumnSelector`

* ColumnSelector init setup and nested subgroup level checks (#1015)

* changes for column_selector init and testing for nested checks

* add typing to kwarg in init func

* fixes from reveiw and some triton fixes need more still

* formatting changes

* fixes for triton

* change arg name to correctly reflect self

* review comments added

* Apply formatting

* Make all `Operator` sub-class return `ColumnSelector` for output names

* Update `Operator` defined in Outbrain example to use `ColumnSelector`

* Restore the `["a", "b", "c"] >> ops.Operator()` syntax

Removing this helped separate `ColumnGroup` into `ColumnSelector` and `WorkflowNode` (the new name of the graph parts of `ColumnGroup` applied in a future PR), but it's still possible to implement this after the big refactor in order to maintain backward compatibility. This commit rolls back changes to the test suite and notebook related to the earlier removal of this syntax, re-implements the behavior, and adds a deprecation warning.

* Roll back `ColumnSelector` related changes in `test_io.py`

* Resolve outstanding `ColumnGroup` TODOs

* Fix deprecation warning text for list rshift syntax

* Add a docstring to `ColumnSelector`

* Add Sphinx docs for `ColumnSelector`

* Add `ColumnSelector` to Sphinx docs index

* Fix underline in `ColumnSelector` docs

* Wrap Triton call to `ColumnGroup.input_column_names` w/ `ColumnSelector`

* Convert list to `ColumnSelector` in `ColumnGroup.columns` setter

Co-authored-by: Julio Perez <37191411+jperez999@users.noreply.github.com>
Co-authored-by: Ben Frederickson <github@benfrederickson.com>