Stop creating a `WorkflowNode` when a `ColumnSelector` will do
`JoinGroupby` was one of the rare cases where an operator sometimes created a `WorkflowNode` for its dependencies, which ended up being an input node with no op when the dependencies were columns only. Now that we've reworked the way dependencies work to avoid needing to create input nodes for column-only dependencies without transformations, this can/should be updated to match.