Add tests for issue #1074 and update Categorify args
Add tests for issue #1074 on adding a start_index argument to
the Categorify op. The motivation for this issue is to allow an
offset for encoding out-of-vocabulary and other special values.

In this commit, we add a test function to test various values of our
new start_index arg in tests/unit/ops/text_ops.py and add a
start_value arg to the class signature for Categorify(),
documentation for its intended function in Categorify()'s
docstring, a start_value arg to FitOptions(), and documentation
for this new argument in FitOptions()'s docstring.