Implement the Categorify() start_index feature request of #1074
This commit implements the feature in issue #1074. This issue
asks to add an argument start_index to Categorify to give an offset
for translating vocabulary items to categorical values.

We update nvtabular/ops/categorify.py to add a start_index arg in the
implementation of Categorify(). This update touches the categorify.py module
in various places. We also add docstrings to the _encode() and
_write_uniques() methods for improved readability in categorify.py.

We also update the test_categorify_lists_with_start_index() test
method in tests/unit/test_ops.py to test various start_index
values.