Update categorify and test_ops modules with pre-commit
Update categorify and test_ops modules with the output of the
pre-commit package. In particular, this updates these files
with the output of various linters like black, flake8, and
pylint, as specified in the .pre-commit-config.yaml file.
The motivation for this change is for better readability and
for to pass the linting steps of the CI pipeline.