Implementation of left padding for issue #1077
Implementation of left padding for issue #1077. This is based on a suggestion
by @gabrielspmoreira. I am not exactly sure if this change will completely
work, and this is untested due to current failing tests on main on this part of
the codebase. But the motivation of this commit is to start a commit for
comments, suggestions, and revisions on this issue's implementation.