Update na_sentinel in categorify
Update na_sentinel in categorify to have a default value of -1.
The motivation of this is to standardize the default value at
either -1 or 0, both of which are currently in use as default
values for na_sentinel. In this commit, we choose -1, since
it seems that 0 is used as a default value for tokens which
fall under the frequency threshold value. The overall motivation
is improved codebase readability and consistency.