Update unit test for padding argument
Update unit test for padding argument for torch dataloader. We add explicit
constructions of our expected output tensors with both padding on the left and
the default padding on the right.  The motivation is to make more clear that
this test is testing these different padding modes for different categorical
columns.