Update dataloader torch to pass new tests
Update dataloader torch to pass new tests for the torch dataloader. We add new
code in order to correctly modify the indices of a sparse matrix in order to
correctly output a dense matrix with padding on the left. Previously the order
of elements in the row were reversed. This new code makes the elements of the
dense matrix in correct order by modifying the indices of its sparse matrix
representation.

We also update pad_left to be a boolean and move this argument to the top-level
TorchAsyncItr() class for better module-level consistency with the rest of the
similar dataloader arguments as pad_left.