Fix conda environment files (#1152)
Our conda environment files were failing to install, because it was overspecifying required
versions of transitive dependencies. For instance, we specified the version of pandas required
but that conflicts with the version that we pick up from cudf (likewise pyarrow etc).

Fix by removing transitive dependencies like pandas/pyarrow etc-  and instead rely on the
fact that these are dependencies of cudf. This means we won't have to keep updating version
numbers ourselves to keep in sync with cudf. Also remove development dependencies, which can
be picked up from requirements-dev.txt instead.