Fix Criteo Inference
The inference notebooks with the Criteo example were broken - and failed
to generate triton config with an in _remove_columns like ```ValueError: list.remove(x): x not in
list```. This was because the label column wa being inserted into a subgroup of the outputnode,
and wasn't getting removed from there.

Fix and add a basic unittest for _remove_columns that would have caught this.

Closes https://github.com/NVIDIA-Merlin/NVTabular/issues/1198