Update Boto3 version (#1116)
Update requirements for issue #1105 to address an issue where
pip needs to sequentially download and verify a long list
of boto3 minor versions. This results in delays to the CI
jobs on the order of six-seven minutes. This commit updates
the boto3 version to a tighter specification to avoid this long
sequential check. The motivation is improved CI jobs and new
user local environment installation times.