Fix inference issue handling multiple nodes with same column output
This fixes an issue in the inference code where trying to gather inputs from the input dataframe that aren't produced by the upstream (e.g. parent) nodes relied on a list of the upstream nodes' outputs, which sometimes may contain duplicate column names. When it tries to remove those from the list of input columns required by the current node, the first attempt to remove the column succeeds, but the second one fails. This fix uses a set instead of a list to avoid duplicates.