Fix inference issues for end-to-end TF example for Transformers4Rec (#1299)
* Fix inference issue handling multiple nodes with same column output

This fixes an issue in the inference code where trying to gather inputs from the input dataframe that aren't produced by the upstream (e.g. parent) nodes relied on a list of the upstream nodes' outputs, which sometimes may contain duplicate column names. When it tries to remove those from the list of input columns required by the current node, the first attempt to remove the column succeeds, but the second one fails. This fix uses a set instead of a list to avoid duplicates.

* Fix a call to `Operator.inference_supports` by adding a `hasattr` check

This fixes a call to `inference_supports` (which only exists for `Workflow` nodes, not for generic graph nodes) by wrapping it with a `hasattr` check. This prevents the method from being called for generic graph operators like `ConcatColumns`.

Co-authored-by: Alberto Alvarez <albertoa@nvidia.com>