Pin `merlin-core` to a specific commit to avoid breaking changes (#1409)
There are some upcoming changes to the package names in merlin-core (e.g. graph->dag) that would break the tests in merlin-models if we continued depending directly on the latest commit in merlin-core. Pinning the version to a specific commit allows us to bump the version here in the same PR that propagates upstream changes from merlin-core without breaking the tests/CI on everyone else's PRs.