Fix install model_config_pb2.py
We are generating the model_config_pb2.py file from a protobuf script,
but doing this as part of the 'build_ext' phase - running at the same
time as building our native extensions. This caused the build to not
pick up the generated python file unless it already existed.

Fix by moving the proto generation to the correct 'build_py' phase