Fix integration tests
With the change to the new unified container, we were pulling in more integration
tests that previously hadn't been included. This changes back to the set of
stable tests that we were using previously.

Adding the remaining integration tests is tracked for 22.07 release as part
of https://github.com/NVIDIA-Merlin/Merlin/issues/343