Fix PyTorch export/serving with Python back-end for T4R notebooks
There's really no reason for us to take the perfectly good 2d tensors that come out of NVT Workflows, convert them to `values`/`offsets` format, pass them to the PyTorch model, and then undo the whole process when we can just pass 2d tensors by setting the dimensions correctly.

(I think this code can be cleaned up and simplified further, but I'm holding off on that in the interest of getting this fix in. In particular, I believe the code that handles the `values`/`offsets` format for PyTorch to be dead at this point, but I want to verify that before removing it.)