Fix `GroupBy` column names when grouping and outputting same column
In this case, the piece of code that should be connecting the column name with the aggregation name incorrectly tries to join every character of the column name with the name separator, resulting in a column called (for example) `c_u_s_t_o_m_e_r___i_d`. We can avoid that by explicitly checking if we have a string (which is iterable like a list) instead of an actual list.