Incorporate `supported_formats` property into NVT ops (#1799)
This adds a property based on a new enumeration in order to provide a clearer definition of the data formats supported by each op. It doesn't use the existing `supports` property/flags, since `CPU` and `GPU` become confusing in the context of data types that can exist on either host or device (like `TensorTable`s backed by Tensorflow and Torch tensors.)

We'll later remove the `supports` property, but keeping it here for now allows us to make this change more incrementally.